Refactor Visual Behavior BehaviorOphysSession and BehaviorSession classes
This commit does a fairly major refactor. Some notable changes include:

1. BehaviorOphysSession now inherits from the BehaviorSession class.
   This reduces amount of duplicated properties/setters and also
   better dilineates the main 'concerns' of the two classes.

2. Use of the AllenSDK LazyPropertyMixin (and class) to allow sessions
   to be compared against each other as well as to slightly reduce some
   boilerplate for checking if attributes were None.

3. Moves get_reward_rate(), get_rolling_performance_df(), and
   get_performance_metrics() methods from BehaviorOphysSession
   to BehaviorSession. These metrics can be calculated with
   just BehaviorSessions attributes/properties.

4. Improvements to docstrings and documentation
